move around to impl new trait for graphvis
