+++ helixdb/src/helix_engine/storage_core/mod.rs
@@ -1,2 +1,4 @@
 pub mod storage_core;
 pub mod storage_methods;
+pub mod graph_visualization;
+
